Initial functioning javascript calculator
JS calculator that accepts user button clicks to build a mathematical
expression in the form of a string, evaluate the expression, and display
the result. The calculator DOES NOT respect mathematical order of
operations, instead evaluating each expression in order from left to
right. Currently, expect for a divide-by-zero check, there is no input
validation. Calculator accepts integer or decimal values.